% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remoji.R
\name{emoji}
\alias{emoji}
\title{Return unicode}
\usage{
emoji(alias, pad = FALSE, must_work = FALSE)
}
\arguments{
\item{alias}{Character vector of emoji aliases.}

\item{pad}{Logical indicating if the returned string should be
padded with a single space?  This seems to help printing.}

\item{must_work}{Logical indicating that a look failure should
cause an error.}
}
\value{
A character vector the same length as \code{alias}
}
\description{
Returns the unicode character for an emoji with given alias.
These cheatsheet at http://emoji-cheat-sheet.com will probably
help, or use \code{\link{find_emoji}} to find an emoji.  The
database comes from https://github.com/github/gemoji
}
\examples{
emoji("+1")
message(emoji("+1"))
message(emoji(c("fire", "computer"), TRUE)[c(1, 1, 1, 2, 1, 1, 1)])
}

