% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remoji.R
\name{emoji_table}
\alias{emoji_table}
\title{Table of emoji}
\usage{
emoji_table(alias)
}
\arguments{
\item{alias}{Vector of aliases}
}
\description{
Print a table of emoji.  May be useful for filtering through to
find the right emoji, especially with \code{\link{find_emoji}}.
}
\examples{
emoji_table(find_emoji("frown", approximate=TRUE))
}

