% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remoji.R
\name{find_emoji}
\alias{find_emoji}
\alias{list_emoji}
\title{Find emoji}
\usage{
find_emoji(str, type = c("alias", "tag", "description"), fixed = FALSE,
  approximate = FALSE)

list_emoji(type = c("alias", "tag", "description"))
}
\arguments{
\item{str}{String to search for}

\item{type}{Field to look in - tag, alias or description}

\item{fixed}{Use fixed string search?  The default is a regular
expression.}

\item{approximate}{Use approximate string matches?}
}
\value{
Vector of possible aliases.  These can be fed into
\code{\link{emoji}} or used in string substitution/interpolation
for \code{sub_emoji}.
}
\description{
Find emoji by alias, tag or full name.  \code{grep} is used for
the match, allowing some flexibility.
}
\examples{
find_emoji("person")
find_emoji("person", "description")
emoji(find_emoji("person", "description"))
message(emoji(find_emoji("person", "description"), pad=TRUE))
}

