% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remoji.R
\name{sub_emoji}
\alias{sub_emoji}
\alias{unsub_emoji}
\title{Emoji string interpolation}
\usage{
sub_emoji(str, pad = TRUE)

unsub_emoji(str, pad = TRUE)
}
\arguments{
\item{str}{A string containing emoji aliases surrounded by
colons.}

\item{pad}{Should each emoji be padded with a single trailing
space?  The default is \code{TRUE} as this looks heaps better here.}
}
\description{
Emoji string interpolation in the style of github.  Intersperse
emoji aliases surrounded by colons with text and the returned
string will replace the aliases with unicode emoji, optionally
padded so that they print nicely.
}

