% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/twemoji.R
\name{twemoji}
\alias{twemoji}
\title{Twitter Emoji (twemoji)}
\usage{
twemoji(svg = FALSE)
}
\arguments{
\item{svg}{\code{logical} to use \code{SVG} instead of \code{PNG}}
}
\description{
Use the Twitter Emoji open-sourced
\href{https://github.com/twitter/twemoji}{twemoji} emoji library
easily in the R console, RStudio Viewer, or \code{rmarkdown}.
}

